pyproject.toml
[project]
name = "mqtt-notifier"
dependencies = [
    "paho-mqtt",
    "argcomplete"
]